Move grouping in diagram, load items in __init__.py
