+++ gaphor/diagram/actions/__init__.py
@@ -1,0 +1,12 @@
+from .action import ActionItem
+from .activitynodes import (
+    ActivityNodeItem,
+    ActivityFinalNodeItem,
+    FlowFinalNodeItem,
+    DecisionNodeItem,
+)
+from .flow import FlowItem
+from .objectnode import ObjectNodeItem
+from .partition import PartitionItem
+
+# from . import actionsgrouping, flowconnect, partitionpage
