+++ gaphor/diagram/classes/__init__.py
@@ -1,0 +1,9 @@
+from .association import AssociationItem
+from .dependency import DependencyItem
+from .generalization import GeneralizationItem
+from .implementation import ImplementationItem
+from .interface import InterfaceItem
+from .klass import ClassItem
+from .package import PackageItem
+
+# from . import classconnect, interfaceconnect
