+++ gaphor/diagram/components/__init__.py
@@ -1,0 +1,7 @@
+from .artifact import ArtifactItem
+from .component import ComponentItem
+from .connector import ConnectorItem
+from .node import NodeItem
+from .subsystem import SubsystemItem
+
+# from . import componentsgrouping, connectorconnect
