+++ gaphor/diagram/diagramitem.py
@@ -9,11 +9,66 @@ import logging
 from gaphor import UML
 from gaphor.services.elementdispatcher import EventWatcher
 from gaphor.core import inject
-from gaphor.diagram import DiagramItemMeta
 from gaphor.diagram.textelement import EditableTextSupport
-from gaphor.diagram.style import ALIGN_CENTER, ALIGN_TOP
+from gaphor.diagram.style import Style, ALIGN_CENTER, ALIGN_TOP
+from .support import set_diagram_item
+
 
 logger = logging.getLogger(__name__)
+
+
+class DiagramItemMeta(type):
+    """
+    Initialize a new diagram item.
+    1. Register UML.Elements by means of the __uml__ attribute (see
+       map_uml_class method).
+    2. Set items style information.
+
+    @ivar style: style information
+    """
+
+    def __init__(self, name, bases, data):
+        type.__init__(self, name, bases, data)
+
+        self.map_uml_class(data)
+        self.set_style(data)
+
+    def map_uml_class(self, data):
+        """
+        Map UML class to diagram item.
+
+        @param cls:  new instance of item class
+        @param data: metaclass data with UML class information
+
+        """
+        if "__uml__" in data:
+            obj = data["__uml__"]
+            if isinstance(obj, (tuple, set, list)):
+                for c in obj:
+                    set_diagram_item(c, self)
+            else:
+                set_diagram_item(obj, self)
+
+    def set_style(self, data):
+        """
+        Set item style information by merging provided information with
+        style information from base classes.
+
+        @param cls:   new instance of diagram item class
+        @param bases: base classes of an item
+        @param data:  metaclass data with style information
+        """
+        style = Style()
+        for c in self.__bases__:
+            if hasattr(c, "style"):
+                for (name, value) in list(c.style.items()):
+                    style.add(name, value)
+
+        if "__style__" in data:
+            for (name, value) in data["__style__"].items():
+                style.add(name, value)
+
+        self.style = style
 
 
 class StereotypeSupport:
