+++ gaphor/diagram/interactions/__init__.py
@@ -1,0 +1,5 @@
+from .interaction import InteractionItem
+from .message import MessageItem
+from .lifeline import LifelineItem
+
+# from . import messageconnect, interactionsgrouping
