+++ gaphor/diagram/profiles/__init__.py
@@ -1,0 +1,4 @@
+from .extension import ExtensionItem
+from .metaclass import MetaclassItem
+
+# from . import extensionconnect, metaclasseditor, stereotypepage
