+++ gaphor/diagram/usecases/__init__.py
@@ -1,0 +1,6 @@
+from .actor import ActorItem
+from .extend import ExtendItem
+from .include import IncludeItem
+from .usecase import UseCaseItem
+
+from . import usecaseconnect
