+++ gaphor/ui/diagrampage.py
@@ -27,7 +27,7 @@ from gaphor.core import (
     action,
     build_action_group,
 )
-from gaphor.diagram import get_diagram_item
+from gaphor.diagram.support import get_diagram_item
 from gaphor.diagram.items import DiagramItem
 from gaphor.transaction import Transaction
 from gaphor.ui.diagramtoolbox import DiagramToolbox
