+++ gaphor/ui/diagramtools.py
@@ -26,7 +26,7 @@ from gi.repository import Gtk
 from gaphor.core import Transaction, transactional
 from gaphor.diagram.diagramline import DiagramLine
 from gaphor.diagram.elementitem import ElementItem
-from gaphor.diagram.interfaces import Group
+from gaphor.diagram.grouping import Group
 from gaphor.diagram.editors import Editor
 from gaphor.diagram.connectors import IConnect
